+++ towerGame/src/main/CollisionChecker.java
@@ -11,17 +11,17 @@ import map.Level;
 import map.Tile;
 
 public class CollisionChecker {
-	public static boolean checkTile(Level level, Entity entity, Direction direction, float movement) {
-		float entityLeftX=entity.x
-				+((float)entity.hitbox.x/16);
-		float entityRightX=entity.x+((float)entity.hitbox.x/16)+((float)entity.hitbox.width/16);
-		float entityTopY=entity.y+((float)entity.hitbox.y/16);
-		float entityBottomY=entity.y+((float)entity.hitbox.y/16)+((float)entity.hitbox.height/16);
+	public static boolean checkTile(Level level, Entity entity, Direction direction, double movement) {
+		double entityLeftX=entity.x
+				+((double)entity.hitbox.x/16);
+		double entityRightX=entity.x+((double)entity.hitbox.x/16)+((double)entity.hitbox.width/16);
+		double entityTopY=entity.y+((double)entity.hitbox.y/16);
+		double entityBottomY=entity.y+((double)entity.hitbox.y/16)+((double)entity.hitbox.height/16);
 		int tileNum1,tileNum2;
-		/*float entityLeftCol=entityLeftX/(Main.tileSize);
-		float entityRightCol=entityLeftX/(Main.tileSize);
-		float entityTopRow=entityLeftX/(Main.tileSize);
-		float entityBottomRow=entityLeftX/(Main.tileSize);*/
+		/*double entityLeftCol=entityLeftX/(Main.tileSize);
+		double entityRightCol=entityLeftX/(Main.tileSize);
+		double entityTopRow=entityLeftX/(Main.tileSize);
+		double entityBottomRow=entityLeftX/(Main.tileSize);*/
 		switch(direction) {
 			case UP:
 				entityTopY-=movement;
@@ -34,22 +34,22 @@ public class CollisionChecker {
 					if(Tile.tiles[tileNum1].hasCustomHitbox||Tile.tiles[tileNum2].hasCustomHitbox) {
 						if(Tile.tiles[tileNum1].isSolid
 								&&(int)entityTopY
-								+((float)Tile.tiles[tileNum1].hitbox.y/16)
-								+((float)Tile.tiles[tileNum1].hitbox.height/16)
+								+((double)Tile.tiles[tileNum1].hitbox.y/16)
+								+((double)Tile.tiles[tileNum1].hitbox.height/16)
 								>entityTopY
 								&&((int)entityLeftX
-								+((float)Tile.tiles[tileNum1].hitbox.x)/16
-								+((float)Tile.tiles[tileNum1].hitbox.width)/16
+								+((double)Tile.tiles[tileNum1].hitbox.x)/16
+								+((double)Tile.tiles[tileNum1].hitbox.width)/16
 								>entityLeftX)) {
 							return true;
 						}
 						if(Tile.tiles[tileNum2].isSolid
 								&&(int)entityTopY
-								+((float)Tile.tiles[tileNum2].hitbox.y/16)
-								+((float)Tile.tiles[tileNum2].hitbox.height/16)
+								+((double)Tile.tiles[tileNum2].hitbox.y/16)
+								+((double)Tile.tiles[tileNum2].hitbox.height/16)
 								>entityTopY
 								&&((int)entityRightX
-								+((float)Tile.tiles[tileNum2].hitbox.x)/16
+								+((double)Tile.tiles[tileNum2].hitbox.x)/16
 								<entityRightX)) {
 							return true;
 						}
@@ -70,20 +70,20 @@ public class CollisionChecker {
 					if(Tile.tiles[tileNum1].hasCustomHitbox||Tile.tiles[tileNum2].hasCustomHitbox) {
 						if(Tile.tiles[tileNum1].isSolid
 								&&(int)entityBottomY
-								+((float)Tile.tiles[tileNum1].hitbox.y/16)
+								+((double)Tile.tiles[tileNum1].hitbox.y/16)
 								<entityBottomY
 								&&((int)entityLeftX
-								+((float)Tile.tiles[tileNum1].hitbox.x)/16
-								+((float)Tile.tiles[tileNum1].hitbox.width)/16
+								+((double)Tile.tiles[tileNum1].hitbox.x)/16
+								+((double)Tile.tiles[tileNum1].hitbox.width)/16
 								>entityLeftX)) {
 							return true;
 						}
 						if(Tile.tiles[tileNum2].isSolid
 								&&(int)entityBottomY
-								+((float)Tile.tiles[tileNum2].hitbox.y/16)
+								+((double)Tile.tiles[tileNum2].hitbox.y/16)
 								<entityBottomY
 								&&((int)entityRightX
-								+((float)Tile.tiles[tileNum2].hitbox.x)/16
+								+((double)Tile.tiles[tileNum2].hitbox.x)/16
 								<entityRightX)) {
 							return true;
 						}
@@ -104,23 +104,23 @@ public class CollisionChecker {
 					if(Tile.tiles[tileNum1].hasCustomHitbox||Tile.tiles[tileNum2].hasCustomHitbox) {
 						if(Tile.tiles[tileNum1].isSolid
 								&&(int)entityLeftX
-								+((float)Tile.tiles[tileNum1].hitbox.x/16)
-								+((float)Tile.tiles[tileNum1].hitbox.width/16)
+								+((double)Tile.tiles[tileNum1].hitbox.x/16)
+								+((double)Tile.tiles[tileNum1].hitbox.width/16)
 								>entityLeftX
 								&&((int)entityBottomY
-								+((float)Tile.tiles[tileNum1].hitbox.y)/16
-								//+((float)Tile.tiles[tileNum1].hitbox.height)/16
+								+((double)Tile.tiles[tileNum1].hitbox.y)/16
+								//+((double)Tile.tiles[tileNum1].hitbox.height)/16
 								<entityBottomY)) {
 							return true;
 						}
 						if(Tile.tiles[tileNum2].isSolid
 								&&(int)entityLeftX
-								+((float)Tile.tiles[tileNum2].hitbox.x/16)
-								+((float)Tile.tiles[tileNum2].hitbox.width/16)
+								+((double)Tile.tiles[tileNum2].hitbox.x/16)
+								+((double)Tile.tiles[tileNum2].hitbox.width/16)
 								>entityLeftX
 								&&((int)entityTopY
-								+((float)Tile.tiles[tileNum2].hitbox.y)/16
-								+((float)Tile.tiles[tileNum2].hitbox.height)/16
+								+((double)Tile.tiles[tileNum2].hitbox.y)/16
+								+((double)Tile.tiles[tileNum2].hitbox.height)/16
 								>entityTopY)) {
 							return true;
 						}
@@ -142,21 +142,21 @@ public class CollisionChecker {
 					if(Tile.tiles[tileNum1].hasCustomHitbox||Tile.tiles[tileNum2].hasCustomHitbox) {
 						if(Tile.tiles[tileNum1].isSolid
 								&&(int)entityRightX
-								+((float)Tile.tiles[tileNum1].hitbox.x/16)
+								+((double)Tile.tiles[tileNum1].hitbox.x/16)
 								<entityRightX
 								&&((int)entityBottomY
-								+((float)Tile.tiles[tileNum1].hitbox.y)/16
-								//+((float)Tile.tiles[tileNum1].hitbox.height)/16
+								+((double)Tile.tiles[tileNum1].hitbox.y)/16
+								//+((double)Tile.tiles[tileNum1].hitbox.height)/16
 								<entityBottomY)) {
 							return true;
 						}
 						if(Tile.tiles[tileNum2].isSolid
 								&&(int)entityRightX
-								+((float)Tile.tiles[tileNum2].hitbox.x/16)
+								+((double)Tile.tiles[tileNum2].hitbox.x/16)
 								<entityRightX
 								&&((int)entityTopY
-								+((float)Tile.tiles[tileNum2].hitbox.y)/16
-								+((float)Tile.tiles[tileNum2].hitbox.height)/16
+								+((double)Tile.tiles[tileNum2].hitbox.y)/16
+								+((double)Tile.tiles[tileNum2].hitbox.height)/16
 								>entityTopY)) {
 							return true;
 						}
@@ -169,17 +169,17 @@ public class CollisionChecker {
 			}
 		return false;
 	}
-	public static boolean checkSpecificTiles(Level level, Entity entity, Direction direction, float movement, List<Tile> tiles) {
-		float entityLeftX=entity.x
-				+((float)entity.hitbox.x/16);
-		float entityRightX=entity.x+((float)entity.hitbox.x/16)+((float)entity.hitbox.width/16);
-		float entityTopY=entity.y+((float)entity.hitbox.y/16);
-		float entityBottomY=entity.y+((float)entity.hitbox.y/16)+((float)entity.hitbox.height/16);
+	public static boolean checkSpecificTiles(Level level, Entity entity, Direction direction, double movement, List<Tile> tiles) {
+		double entityLeftX=entity.x
+				+((double)entity.hitbox.x/16);
+		double entityRightX=entity.x+((double)entity.hitbox.x/16)+((double)entity.hitbox.width/16);
+		double entityTopY=entity.y+((double)entity.hitbox.y/16);
+		double entityBottomY=entity.y+((double)entity.hitbox.y/16)+((double)entity.hitbox.height/16);
 		int tileNum1,tileNum2;
-		/*float entityLeftCol=entityLeftX/(Main.tileSize);
-		float entityRightCol=entityLeftX/(Main.tileSize);
-		float entityTopRow=entityLeftX/(Main.tileSize);
-		float entityBottomRow=entityLeftX/(Main.tileSize);*/
+		/*double entityLeftCol=entityLeftX/(Main.tileSize);
+		double entityRightCol=entityLeftX/(Main.tileSize);
+		double entityTopRow=entityLeftX/(Main.tileSize);
+		double entityBottomRow=entityLeftX/(Main.tileSize);*/
 		switch(direction) {
 			case UP:
 				entityTopY-=movement;
@@ -192,22 +192,22 @@ public class CollisionChecker {
 					if(Tile.tiles[tileNum1].hasCustomHitbox||Tile.tiles[tileNum2].hasCustomHitbox) {
 						if(tiles.contains(Tile.tiles[tileNum1])
 								&&(int)entityTopY
-								+((float)Tile.tiles[tileNum1].hitbox.y/16)
-								+((float)Tile.tiles[tileNum1].hitbox.height/16)
+								+((double)Tile.tiles[tileNum1].hitbox.y/16)
+								+((double)Tile.tiles[tileNum1].hitbox.height/16)
 								>entityTopY
 								&&((int)entityLeftX
-								+((float)Tile.tiles[tileNum1].hitbox.x)/16
-								+((float)Tile.tiles[tileNum1].hitbox.width)/16
+								+((double)Tile.tiles[tileNum1].hitbox.x)/16
+								+((double)Tile.tiles[tileNum1].hitbox.width)/16
 								>entityLeftX)) {
 							return true;
 						}
 						if(tiles.contains(Tile.tiles[tileNum2])
 								&&(int)entityTopY
-								+((float)Tile.tiles[tileNum2].hitbox.y/16)
-								+((float)Tile.tiles[tileNum2].hitbox.height/16)
+								+((double)Tile.tiles[tileNum2].hitbox.y/16)
+								+((double)Tile.tiles[tileNum2].hitbox.height/16)
 								>entityTopY
 								&&((int)entityRightX
-								+((float)Tile.tiles[tileNum2].hitbox.x)/16
+								+((double)Tile.tiles[tileNum2].hitbox.x)/16
 								<entityRightX)) {
 							return true;
 						}
@@ -228,20 +228,20 @@ public class CollisionChecker {
 					if(Tile.tiles[tileNum1].hasCustomHitbox||Tile.tiles[tileNum2].hasCustomHitbox) {
 						if(tiles.contains(Tile.tiles[tileNum1])
 								&&(int)entityBottomY
-								+((float)Tile.tiles[tileNum1].hitbox.y/16)
+								+((double)Tile.tiles[tileNum1].hitbox.y/16)
 								<entityBottomY
 								&&((int)entityLeftX
-								+((float)Tile.tiles[tileNum1].hitbox.x)/16
-								+((float)Tile.tiles[tileNum1].hitbox.width)/16
+								+((double)Tile.tiles[tileNum1].hitbox.x)/16
+								+((double)Tile.tiles[tileNum1].hitbox.width)/16
 								>entityLeftX)) {
 							return true;
 						}
 						if(tiles.contains(Tile.tiles[tileNum2])
 								&&(int)entityBottomY
-								+((float)Tile.tiles[tileNum2].hitbox.y/16)
+								+((double)Tile.tiles[tileNum2].hitbox.y/16)
 								<entityBottomY
 								&&((int)entityRightX
-								+((float)Tile.tiles[tileNum2].hitbox.x)/16
+								+((double)Tile.tiles[tileNum2].hitbox.x)/16
 								<entityRightX)) {
 							return true;
 						}
@@ -262,23 +262,23 @@ public class CollisionChecker {
 					if(Tile.tiles[tileNum1].hasCustomHitbox||Tile.tiles[tileNum2].hasCustomHitbox) {
 						if(tiles.contains(Tile.tiles[tileNum1])
 								&&(int)entityLeftX
-								+((float)Tile.tiles[tileNum1].hitbox.x/16)
-								+((float)Tile.tiles[tileNum1].hitbox.width/16)
+								+((double)Tile.tiles[tileNum1].hitbox.x/16)
+								+((double)Tile.tiles[tileNum1].hitbox.width/16)
 								>entityLeftX
 								&&((int)entityBottomY
-								+((float)Tile.tiles[tileNum1].hitbox.y)/16
-								//+((float)Tile.tiles[tileNum1].hitbox.height)/16
+								+((double)Tile.tiles[tileNum1].hitbox.y)/16
+								//+((double)Tile.tiles[tileNum1].hitbox.height)/16
 								<entityBottomY)) {
 							return true;
 						}
 						if(tiles.contains(Tile.tiles[tileNum2])
 								&&(int)entityLeftX
-								+((float)Tile.tiles[tileNum2].hitbox.x/16)
-								+((float)Tile.tiles[tileNum2].hitbox.width/16)
+								+((double)Tile.tiles[tileNum2].hitbox.x/16)
+								+((double)Tile.tiles[tileNum2].hitbox.width/16)
 								>entityLeftX
 								&&((int)entityTopY
-								+((float)Tile.tiles[tileNum2].hitbox.y)/16
-								+((float)Tile.tiles[tileNum2].hitbox.height)/16
+								+((double)Tile.tiles[tileNum2].hitbox.y)/16
+								+((double)Tile.tiles[tileNum2].hitbox.height)/16
 								>entityTopY)) {
 							return true;
 						}
@@ -300,21 +300,21 @@ public class CollisionChecker {
 					if(Tile.tiles[tileNum1].hasCustomHitbox||Tile.tiles[tileNum2].hasCustomHitbox) {
 						if(tiles.contains(Tile.tiles[tileNum1])
 								&&(int)entityRightX
-								+((float)Tile.tiles[tileNum1].hitbox.x/16)
+								+((double)Tile.tiles[tileNum1].hitbox.x/16)
 								<entityRightX
 								&&((int)entityBottomY
-								+((float)Tile.tiles[tileNum1].hitbox.y)/16
-								//+((float)Tile.tiles[tileNum1].hitbox.height)/16
+								+((double)Tile.tiles[tileNum1].hitbox.y)/16
+								//+((double)Tile.tiles[tileNum1].hitbox.height)/16
 								<entityBottomY)) {
 							return true;
 						}
 						if(tiles.contains(Tile.tiles[tileNum2])
 								&&(int)entityRightX
-								+((float)Tile.tiles[tileNum2].hitbox.x/16)
+								+((double)Tile.tiles[tileNum2].hitbox.x/16)
 								<entityRightX
 								&&((int)entityTopY
-								+((float)Tile.tiles[tileNum2].hitbox.y)/16
-								+((float)Tile.tiles[tileNum2].hitbox.height)/16
+								+((double)Tile.tiles[tileNum2].hitbox.y)/16
+								+((double)Tile.tiles[tileNum2].hitbox.height)/16
 								>entityTopY)) {
 							return true;
 						}
@@ -327,30 +327,30 @@ public class CollisionChecker {
 			}
 		return false;
 	}
-	public static boolean checkSpecificTile(Level level, Entity entity, Direction direction, float movement, Tile tile) {
+	public static boolean checkSpecificTile(Level level, Entity entity, Direction direction, double movement, Tile tile) {
 		List<Tile> list = new ArrayList<Tile>();
 		list.add(tile);
 		return checkSpecificTiles(level, entity, direction, movement, list);
 	}
 		
-	public static int[] getTilePositions(Level level, Entity entity, Direction direction, float movement) {
-		float entityLeftX=entity.x
-				+((float)entity.hitbox.x/16);
-		float entityRightX=entity.x+((float)entity.hitbox.x/16)+((float)entity.hitbox.width/16);
-		float entityTopY=entity.y+((float)entity.hitbox.y/16);
-		float entityBottomY=entity.y+((float)entity.hitbox.y/16)+((float)entity.hitbox.height/16);
+	public static int[] getTilePositions(Level level, Entity entity, Direction direction, double movement) {
+		double entityLeftX=entity.x
+				+((double)entity.hitbox.x/16);
+		double entityRightX=entity.x+((double)entity.hitbox.x/16)+((double)entity.hitbox.width/16);
+		double entityTopY=entity.y+((double)entity.hitbox.y/16);
+		double entityBottomY=entity.y+((double)entity.hitbox.y/16)+((double)entity.hitbox.height/16);
 		switch(direction) {
 		case UP:
-			entityTopY=entity.y+((float)entity.hitbox.y/16)-movement;
+			entityTopY=entity.y+((double)entity.hitbox.y/16)-movement;
 			break;
 		case DOWN:
-			entityBottomY=entity.y+((float)entity.hitbox.y/16)+((float)entity.hitbox.height/16)+movement;
+			entityBottomY=entity.y+((double)entity.hitbox.y/16)+((double)entity.hitbox.height/16)+movement;
 			break;
 		case LEFT:
-			entityLeftX=entity.x+((float)entity.hitbox.x/16)-movement;
+			entityLeftX=entity.x+((double)entity.hitbox.x/16)-movement;
 			break;
 		case RIGHT:
-			entityRightX=entity.x+((float)entity.hitbox.x/16)+((float)entity.hitbox.width/16)+movement;
+			entityRightX=entity.x+((double)entity.hitbox.x/16)+((double)entity.hitbox.width/16)+movement;
 			break;
 		}
 		int[] positions={(int)entityLeftX,(int)entityRightX,(int)entityTopY,(int)entityBottomY};
@@ -358,10 +358,10 @@ public class CollisionChecker {
 	}
 	public static void renderDebug(Level level,Entity entity,Graphics2D g2) {
 		g2.setColor(new Color(192,0,0,64));
-		float entityLeftX=entity.x+((float)entity.hitbox.x/16);
-		float entityRightX=entity.x+((float)entity.hitbox.x/16)+((float)entity.hitbox.width/16);
-		float entityTopY=entity.y+((float)entity.hitbox.y/16);
-		float entityBottomY=entity.y+((float)entity.hitbox.y/16)+((float)entity.hitbox.height/16);
+		double entityLeftX=entity.x+((double)entity.hitbox.x/16);
+		double entityRightX=entity.x+((double)entity.hitbox.x/16)+((double)entity.hitbox.width/16);
+		double entityTopY=entity.y+((double)entity.hitbox.y/16);
+		double entityBottomY=entity.y+((double)entity.hitbox.y/16)+((double)entity.hitbox.height/16);
 		g2.fillRect(((int)((int)entityLeftX*Main.tileSize))-(int)(level.cameraX*Main.tileSize),((int)((int)entityBottomY*Main.tileSize))-(int)(level.cameraY*Main.tileSize),Main.tileSize,Main.tileSize);
 		if((int)entityLeftX!=(int)entityRightX) {
 			g2.fillRect(((int)((int)entityRightX*Main.tileSize))-(int)(level.cameraX*Main.tileSize),((int)((int)entityBottomY*Main.tileSize))-(int)(level.cameraY*Main.tileSize),Main.tileSize,Main.tileSize);
@@ -376,11 +376,11 @@ public class CollisionChecker {
 	public static Rectangle getHitbox(int x0,int y0,int x1,int y1) {
 		return new Rectangle(x0,y0,x1-x0,y1-y0);
 	}
-	public static boolean checkAABB(float x0, float y0, float x1, float y1, float x2, float y2, float x3, float y3) {
+	public static boolean checkAABB(double x0, double y0, double x1, double y1, double x2, double y2, double x3, double y3) {
 		return (x0<=x3)&&(x1>=x2)&&(y0<=y3)&&(y1>=y2);
 	}
-	public static boolean checkHitboxes(Rectangle h1, Rectangle h2, float h1x, float h1y, float h2x, float h2y) {
-		return checkAABB(h1x+((float)h1.x/16), h1y+((float)h1.y/16), h1x+((float)h1.x/16)+((float)h1.width/16), h1y+((float)h1.y/16)+((float)h1.height/16), h2x+((float)h2.x/16), h2y+((float)h2.y/16), h2x+((float)h2.x/16)+((float)h2.width/16), h2y+((float)h2.y/16)+((float)h2.height/16));
+	public static boolean checkHitboxes(Rectangle h1, Rectangle h2, double h1x, double h1y, double h2x, double h2y) {
+		return checkAABB(h1x+((double)h1.x/16), h1y+((double)h1.y/16), h1x+((double)h1.x/16)+((double)h1.width/16), h1y+((double)h1.y/16)+((double)h1.height/16), h2x+((double)h2.x/16), h2y+((double)h2.y/16), h2x+((double)h2.x/16)+((double)h2.width/16), h2y+((double)h2.y/16)+((double)h2.height/16));
 	}
 	public static boolean checkEntities(Entity e1, Entity e2) {
 		return checkHitboxes(e1.hitbox,e2.hitbox,e1.x,e1.y,e2.x,e2.y);
