+++ towerGame/src/weapon/Staff.java
@@ -19,10 +19,10 @@ public class Staff extends Weapon {
 		if(isMouseRight) {
 			if(player.mana>=1) {
 				Point mousePos= MouseInfo.getPointerInfo().getLocation();
-				float angle=(float)Math.atan2((mousePos.x-TowerGame.gamePanel.frame.getLocation().x)-Math.round(player.x*Main.tileSize-(int)(level.cameraX*Main.tileSize)+0.5*Main.tileSize), (mousePos.y-TowerGame.gamePanel.frame.getLocation().y)-Math.round(player.y*Main.tileSize-(int)(level.cameraY*Main.tileSize)+0.5*Main.tileSize));
+				double angle=(double)Math.atan2((mousePos.x-TowerGame.gamePanel.frame.getLocation().x)-Math.round(player.x*Main.tileSize-(int)(level.cameraX*Main.tileSize)+0.5*Main.tileSize), (mousePos.y-TowerGame.gamePanel.frame.getLocation().y)-Math.round(player.y*Main.tileSize-(int)(level.cameraY*Main.tileSize)+0.5*Main.tileSize));
 				PlayerProjectile p = new PlayerProjectile(level, player);
-				p.xVelocity=(float) Math.sin(angle)/5;
-				p.yVelocity=(float) (Math.cos(angle)/5)-0.1F;
+				p.xVelocity=(double) Math.sin(angle)/5;
+				p.yVelocity=(double) (Math.cos(angle)/5)-0.1F;
 				p.size=particleSize+2;
 				level.addEntity(p);
 				player.mana -= 1F;
@@ -31,10 +31,10 @@ public class Staff extends Weapon {
 		}else {
 			if(player.mana>=0.1) {
 				Point mousePos= MouseInfo.getPointerInfo().getLocation();
-				float angle=(float)Math.atan2((mousePos.x-TowerGame.gamePanel.frame.getLocation().x)-Math.round(player.x*Main.tileSize-(int)(level.cameraX*Main.tileSize)+0.5*Main.tileSize), (mousePos.y-TowerGame.gamePanel.frame.getLocation().y)-Math.round(player.y*Main.tileSize-(int)(level.cameraY*Main.tileSize)+0.5*Main.tileSize));
+				double angle=(double)Math.atan2((mousePos.x-TowerGame.gamePanel.frame.getLocation().x)-Math.round(player.x*Main.tileSize-(int)(level.cameraX*Main.tileSize)+0.5*Main.tileSize), (mousePos.y-TowerGame.gamePanel.frame.getLocation().y)-Math.round(player.y*Main.tileSize-(int)(level.cameraY*Main.tileSize)+0.5*Main.tileSize));
 				PlayerProjectile p = new PlayerProjectile(level, player);
-				p.xVelocity=(float) Math.sin(angle)/5;
-				p.yVelocity=(float) (Math.cos(angle)/5)-0.1F;
+				p.xVelocity=(double) Math.sin(angle)/5;
+				p.yVelocity=(double) (Math.cos(angle)/5)-0.1F;
 				p.size=particleSize;
 				level.addEntity(p);
 				player.mana -= 0.1F;
